eval_framework/baselines/MIRIX/requirements.txt
pytz
numpy
pandas
openpyxl
Markdown
langchain_core
pyautogui
Pillow
scikit-image
PyQt6
openai
tiktoken
google-genai
dotenv
demjson3
pathvalidate
docstring_parser
sqlalchemy
pydantic-settings
jinja2
humps
colorama
anthropic
httpx_sse
Markdown
rapidfuzz
rank-bm25
psutil
llama_index
llama-index-embeddings-google-genai
fastapi==0.104.1
uvicorn[standard]==0.24.0
SpeechRecognition
pydub
python-multipart
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests
ffmpeg
pg8000
pgvector
json_repair